Working out which Pathfinders validation constants are likely to persist between rounds and which are likely to be round-specific and moving around accordingly
